"Fix Testcontainers and Readme.md"

+++ src/main/java/com/akul/microservices/product/dto/ProductResponse.java
@@ -11,7 +11,7 @@ import java.math.BigDecimal;
  * @version 1.0
  * @since 8/19/2025
  */
-public record ProductResponse(String id, String name, String description, Double price) {
+public record ProductResponse(String id, String name, String description, BigDecimal price) {
     public static ProductResponse from(Product product) {
         return new ProductResponse(
                 product.getId(),
